+++ spec/ruby/core/io/readlines_spec.rb
@@ -217,10 +217,10 @@ describe "IO.readlines" do
 
   it "encodes lines using the default internal encoding, when set" do
     Encoding.default_external = Encoding::UTF_8
-    Encoding.default_internal = Encoding::UTF_16
-    suppress_warning {$/ = $/.encode Encoding::UTF_16}
+    Encoding.default_internal = Encoding::UTF_16BE
+    suppress_warning {$/ = $/.encode Encoding::UTF_16BE}
     lines = IO.readlines(@name)
-    lines.all? { |s| s.encoding == Encoding::UTF_16 }.should be_true
+    lines.all? { |s| s.encoding == Encoding::UTF_16BE }.should be_true
   end
 
   it "ignores the default internal encoding if the external encoding is BINARY" do