+++ spec/ruby/core/marshal/dump_spec.rb
@@ -61,8 +61,6 @@ describe "Marshal.dump" do
       s = "\u2192"
       [ [Marshal, s.encode("utf-8").to_sym,
             "\x04\bI:\b\xE2\x86\x92\x06:\x06ET"],
-        [Marshal, s.encode("utf-16").to_sym,
-            "\x04\bI:\t\xFE\xFF!\x92\x06:\rencoding\"\vUTF-16"],
         [Marshal, s.encode("utf-16le").to_sym,
             "\x04\bI:\a\x92!\x06:\rencoding\"\rUTF-16LE"],
         [Marshal, s.encode("utf-16be").to_sym,